Remain tests. Javadocs added and some service and localization have been fixed

+++ Keddit/src/main/java/by/parakhnevich/keddit/controller/command/action/edit/DenyPublicationCommand.java
@@ -1,6 +1,7 @@
 package by.parakhnevich.keddit.controller.command.action.edit;
 
 import by.parakhnevich.keddit.bean.publication.Publication;
+import by.parakhnevich.keddit.bean.user.Role;
 import by.parakhnevich.keddit.bean.user.User;
 import by.parakhnevich.keddit.controller.command.Command;
 import by.parakhnevich.keddit.controller.command.CommandPage;
@@ -9,29 +10,33 @@ import by.parakhnevich.keddit.service.ServiceFactory;
 import by.parakhnevich.keddit.service.exception.ServiceException;
 import by.parakhnevich.keddit.service.interfaces.PublicationService;
 import by.parakhnevich.keddit.service.interfaces.UserService;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
+/**
+ * The class DenyPublicationCommand that is Command for
+ * Controller Pattern.
+ * @see Command
+ * @see by.parakhnevich.keddit.controller.command.CommandProvider
+ * @see by.parakhnevich.keddit.controller.KedditController
+ * @author Danila Parakhnevich
+ */
 public class DenyPublicationCommand implements Command {
-    Logger logger = LogManager.getLogger(DenyPublicationCommand.class);
     @Override
-    public void execute(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException, ServiceException {
-        try {
-            UserService userService = ServiceFactory.getInstance().getUserService();
-            User user = userService.selectById(((User)request.getSession().getAttribute("user")).getId());
-            PublicationService publicationService = ServiceFactory.getInstance().getPublicationService();
-            Publication publication = publicationService.selectById(Long.parseLong(request.getParameter("id")));
-            publicationService.delete(publication);
-            request.getSession().setAttribute("user", user);
-            response.sendRedirect("/keddit.by/controller?command=on_moderation_page");
-        } catch (PersistentException e) {
-            logger.error(e);
+    public void execute(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException, ServiceException, PersistentException {
+        UserService userService = ServiceFactory.getInstance().getUserService();
+        User user = userService.selectById(((User)request.getSession().getAttribute("user")).getId());
+        if (user.getRole() == Role.USER) {
             request.getRequestDispatcher(CommandPage.ERROR_PAGE).forward(request, response);
+            return;
         }
+        PublicationService publicationService = ServiceFactory.getInstance().getPublicationService();
+        Publication publication = publicationService.selectById(Long.parseLong(request.getParameter("id")));
+        publicationService.delete(publication);
+        request.getSession().setAttribute("user", user);
+        response.sendRedirect("/keddit.by/controller?command=on_moderation_page");
     }
 }
